refactor: Move cf-environment.ts into `utils`
